fix(fe): change component name
